Abfragen beim löschen der Tasks und Tasklisten

+++ de.nordakademie.wpk.tasklist.ui/src/de/nordakademie/wpk/tasklist/ui/handler/DeleteTasklistHandler.java
@@ -1,13 +1,17 @@
  
 package de.nordakademie.wpk.tasklist.ui.handler;
 
+import java.util.concurrent.ExecutionException;
+
 import javax.inject.Inject;
 
 import org.eclipse.e4.core.di.annotations.Execute;
 import org.eclipse.e4.core.di.annotations.CanExecute;
 import org.eclipse.e4.core.services.events.IEventBroker;
 import org.eclipse.e4.ui.workbench.modeling.ESelectionService;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.ITreeSelection;
+import org.eclipse.swt.widgets.Shell;
 
 import de.nordakademie.wpk.tasklist.core.api.Task;
 import de.nordakademie.wpk.tasklist.core.api.TaskList;
@@ -25,13 +29,18 @@ public class DeleteTasklistHandler {
 	
 	@Inject
 	private TaskService taskService;
+
+	private String message = "Wollen Sie die Taskliste wirklich löschen? Alle Tasks in der Taskliste werden mit gelöscht!";
 	
 	@Execute
-	public void execute() {
+	public void execute(Shell shell ) throws ExecutionException{
 		if (canExecute()) {
+			boolean confirmed = MessageDialog.openQuestion(shell, "Wirklich?", message );
+			if (confirmed) {
 			ITreeSelection selection = (ITreeSelection) selectionService.getSelection();
 			TaskList tasklist = (TaskList) selection.getFirstElement();
 			new DeleteTasklistJob(tasklist.getId(), taskService, eventBroker).schedule();
+			}
 		}
 	}
